Fixes instructions
Solves #1 

Thank you for flagging this, @wtennis! You're always on top of it and we appreciate your contributions! This lab builds on the other lessons from this section, and that was our mistake for not conveying that in the readme (and for pretty much saying otherwise!)
I just updated the ReadME to better reflect what needs to be done in this lab.
Thanks again!